Move GoogleCloudCredentialsHook to airflow.hooks
